Add files via upload
This is my business project that i am planning to built as my first enterprise .This is a fully functional full-stack web application that is aimed to provide the niche of 'waste management' in the domain of "sustainable development". this is a marketplace for a very practical need that is waste management by individuals and also by industries.This website is built by me with the help of my brother.